+++ db/migrate/20170415175000_add_suite_selection_mode_to_draw.rb
@@ -1,0 +1,5 @@
+class AddSuiteSelectionModeToDraw < ActiveRecord::Migration[5.0]
+  def change
+    add_column :draws, :suite_selection_mode, :integer, null: false, default: 0
+  end
+end